refactor(22182): relocated RJSF custom templates for reuse
