+++ hivemq-edge/src/frontend/src/components/rjsf/ArrayFieldTemplate.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import { ArrayFieldTemplateProps } from '@rjsf/utils'
 import { Box, Grid, GridItem } from '@chakra-ui/react'
 import { getTemplate, getUiOptions, ArrayFieldTemplateItemType } from '@rjsf/utils'
-import AddButton from '@/modules/ProtocolAdapters/components/templates/__internals/AddButton.tsx'
+import AddButton from '@/components/rjsf/__internals/AddButton.tsx'
 
 export const ArrayFieldTemplate: FC<ArrayFieldTemplateProps> = (props) => {
   const { canAdd, disabled, idSchema, uiSchema, items, onAddClick, readonly, registry, required, schema, title } = props
