+++ hivemq-edge/src/frontend/src/components/rjsf/ObjectFieldTemplate.tsx
